requirements.txt
Django==1.11.4
tweepy==3.5.0
django-picklefield==1.0.0
